+++ src/include/hud.h
@@ -17,13 +17,13 @@
 #define IS_TOOL_BALL(tool) ((tool) >= TOOL_BALL_0 && (tool) < TOOL_BALL_0 + 8)
 
 typedef struct lb_hud {
-    size_t tool, toolbox, map, stage_b; 
+    size_t tool, toolbox, map, stage_b;
     SDL_Texture* frame;
     SDL_Texture* tools;
     SDL_Texture* balls;
 } lb_hud;
 
-size_t in_rect(const SDL_Rect* r, ssize_t x, ssize_t y);
+size_t in_rect(const SDL_Rect* r, int x, int y);
 lb_hud* hud_init(SDL_Renderer* renderer, lb_sprites* sprites);
 void hud_destroy(lb_hud* hud);
 void hud_click(lb_hud* hud, size_t x, size_t y, size_t* invalid_layers);
